requirements.txt
faster-whisper>=1.0.0
sounddevice>=0.4.6
keyboard>=0.13.5
pyperclip>=1.8.2
requests>=2.31.0
pystray>=0.19.5
pillow>=10.2.0
numpy>=1.26.0
pyyaml>=6.0.1
pywin32>=306; sys_platform == "win32"